adding average back to the model

+++ src/main/java/com/itible/rpmbackend/dto/TrainingDto.java
@@ -2,7 +2,6 @@ package com.itible.rpmbackend.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -11,6 +10,6 @@ public class TrainingDto {
     private Date date;
     private List<Integer> rpm;
     private Integer duration;
-    private BigDecimal average;
+    private Double average;
     private String personName;
 }
